requirements.txt
RPi.GPIO==0.7.1
smbus==1.1.post2
python-dotenv==1.0.0
pinyin==0.4.0
PyGObject==3.44.1
dbus-python==1.3.2
rpi-ws281x==4.3.4
